requirements.txt
smolagents[litellm]
requests
python-dotenv 
duckduckgo-search